third_party/WebKit/LayoutTests/fast/forms/date/no-page-popup-controller.html

<!DOCTYPE html>
<body>
<script src="../../../resources/js-test.js"></script>
<input type=date>
<script>
description('Confirm that the main window doesn\'t have pagePopupController, which should be present only in a calendar picker.');
shouldBeTrue('window.pagePopupController === undefined');
</script>
</body>
